Refactor header and created new separate component searchbar
